+++ packages/analytics-react-native/test/migration/remnant-data-migration.test.ts
@@ -1,7 +1,6 @@
 import { NativeModules } from 'react-native';
 import { AmplitudeReactNative } from '../../src/react-native-client';
-import { Logger, MemoryStorage } from '@amplitude/analytics-core';
-import {
+import { Logger, MemoryStorage ,
   Event,
   LogLevel,
   UserSession,