+++ packages/analytics-react-native/test/mock/setup-mobile.ts
@@ -1,6 +1,5 @@
 import mockAsyncStorage from '@react-native-async-storage/async-storage/jest/async-storage-mock';
-import { NativeModules } from 'react-native';
-import { Platform } from 'react-native';
+import { NativeModules , Platform } from 'react-native';
 
 /*
  * Set the platform OS to mobile.