+++ packages/analytics-node/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.5.20](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.19...@amplitude/analytics-node@1.5.20) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
+## [1.5.19](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.18...@amplitude/analytics-node@1.5.19) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/analytics-node
+
+
+
+
+
 ## [1.5.18](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/analytics-node@1.5.17...@amplitude/analytics-node@1.5.18) (2025-10-17)
 
 **Note:** Version bump only for package @amplitude/analytics-node