+++ packages/plugin-network-capture-browser/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.6.11](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.6.10...@amplitude/plugin-network-capture-browser@1.6.11) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/plugin-network-capture-browser
+
+
+
+
+
+## [1.6.10](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.6.9...@amplitude/plugin-network-capture-browser@1.6.10) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/plugin-network-capture-browser
+
+
+
+
+
 ## [1.6.9](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/plugin-network-capture-browser@1.6.8...@amplitude/plugin-network-capture-browser@1.6.9) (2025-10-17)
 
 **Note:** Version bump only for package @amplitude/plugin-network-capture-browser