+++ packages/unified/CHANGELOG.md
@@ -3,6 +3,33 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+# [1.0.0-beta.13](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified@1.0.0-beta.12...@amplitude/unified@1.0.0-beta.13) (2025-10-23)
+
+**Note:** Version bump only for package @amplitude/unified
+
+
+
+
+
+# [1.0.0-beta.12](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified@1.0.0-beta.11...@amplitude/unified@1.0.0-beta.12) (2025-10-23)
+
+
+### Bug Fixes
+
+* **unified:** export Identify and Revenue ([#1363](https://github.com/amplitude/Amplitude-TypeScript/issues/1363)) ([a47b8c9](https://github.com/amplitude/Amplitude-TypeScript/commit/a47b8c9fb0f59567ef7afb00f70aede885d949d1))
+
+
+
+
+
+# [1.0.0-beta.11](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified@1.0.0-beta.10...@amplitude/unified@1.0.0-beta.11) (2025-10-21)
+
+**Note:** Version bump only for package @amplitude/unified
+
+
+
+
+
 # [1.0.0-beta.10](https://github.com/amplitude/Amplitude-TypeScript/compare/@amplitude/unified@1.0.0-beta.9...@amplitude/unified@1.0.0-beta.10) (2025-10-17)
 
 **Note:** Version bump only for package @amplitude/unified